+++ tests/canonical/test_grading_composite_state_checks.py
@@ -6,7 +6,7 @@ an :class:`InProcessGradingSubstrate` over a hand-built ``{tables: [rows],
 filesystem: {rel: text}}`` fixture, drives the composite through the
 shipping ``jsonpath`` + ``db_probes`` backends, and asserts the
 ``(jsonpath_score, jsonpath_reasons)`` pair matches what
-:func:`evaluate_jsonpath_checks` produces over the same reshaped state
+:func:`~tolokaforge.core.grading.jsonpath_evaluators.evaluate_jsonpath_checks` produces over the same reshaped state
 (``{db, tables, filesystem}``). A drift in the ``jsonpath`` backend's
 reshaping / gating / STABLE routing surfaces here.
 
@@ -29,10 +29,10 @@ from unittest.mock import MagicMock
 import pytest
 
 from tolokaforge.core.grading.composite import grade_state_checks_reads
+from tolokaforge.core.grading.jsonpath_evaluators import evaluate_jsonpath_checks
 from tolokaforge.core.grading.substrate import InProcessGradingSubstrate
 from tolokaforge.core.logging import StructuredLogger
 from tolokaforge.core.plugin_registry import load_state_check_backend
-from tolokaforge.runner.grading import evaluate_jsonpath_checks
 from tolokaforge.runner.grading_ledger import DB_PROBES_KEY, JSONPATHS_KEY
 from tolokaforge.runner.models import RunnerStateChecksConfig
 
@@ -172,8 +172,8 @@ class TestJsonpathCompositeParity:
 
     def test_path_glob_only_pack_matches_the_runner_path(self, tmp_path) -> None:
         # Write a file so the glob check can find something on disk. The
-        # runner's ``evaluate_jsonpath_file_checks`` reads from cwd; setting
-        # cwd via monkeypatch keeps the shipped shape.
+        # ``core.grading.jsonpath_evaluators`` file evaluator reads from cwd;
+        # setting cwd via monkeypatch keeps the shipped shape.
         target = tmp_path / "output.txt"
         target.write_text("hello disk", encoding="utf-8")
         # A pack whose only assertion is a path_glob does not touch the
