+++ tolokaforge/runner/grading.py
@@ -1,26 +1,26 @@
+"""Runner-wire helpers around GradeTrial's state-diff family.
+
+Two responsibilities live here — both runner-wire adjacent, both consumed
+by ``runner.service`` on the GradeTrial path:
+
+- :func:`compute_state_diff` (+ private helpers) — human-readable diff
+  between the trial's final stable state and the golden state, shaped
+  into the runner-side :class:`~tolokaforge.runner.models.StateDiff` /
+  :class:`~tolokaforge.runner.models.TableDiff` types.
+- :func:`project_state_checks_to_runner_wire` — encode the composite
+  fold's neutral ``None``-means-not-evaluated ``state_checks`` slot into
+  the runner wire's ``-1.0`` sentinel.
+
+See docs/GRPC_PROTOCOL.md for the grading algorithm specification. The
+JSONPath assertion evaluators and the SQL-probe evaluator live in
+:mod:`tolokaforge.core.grading.jsonpath_evaluators` and
+:mod:`tolokaforge.core.grading.db_probes` — pure grading libraries the
+runner-side GradeTrial and the standalone Grader v3 service both drive.
 """
-Grading Helper Functions for Runner Service
 
-This module provides helper functions for the GradeTrial RPC implementation:
-- compute_state_diff: Compute human-readable diff between two stable states
-- evaluate_jsonpath_checks / evaluate_jsonpath_file_checks / evaluate_jsonpath_state_checks:
-  runner-side JSONPath assertion evaluators
-- evaluate_db_probes: substrate SQL probes against a task-declared postgres
-- project_state_checks_to_runner_wire: encode the composite fold's neutral
-  ``None``-means-not-evaluated ``state_checks`` slot into the runner wire's
-  ``-1.0`` sentinel
-
-See docs/GRPC_PROTOCOL.md for grading algorithm specification.
-"""
-
-import glob
 import logging
-from pathlib import Path
 from typing import Any
 
-from jsonpath_ng.ext import parse
-
-from tolokaforge.core.grading.predicates import contains
 from tolokaforge.core.hash import canonical_number
 from tolokaforge.runner.models import (
     StateDiff,
@@ -202,309 +202,6 @@ def _get_field_diffs(expected: dict[str, Any], actual: dict[str, Any]) -> list[d
             diffs.append({"field": field, "expected": exp_val, "actual": act_val})
 
     return diffs
-
-
-def evaluate_jsonpath_file_checks(
-    checks: list[dict[str, Any]],
-) -> tuple[float, str]:
-    """
-    Evaluate jsonpath file assertions against the Runner container's filesystem.
-
-    This evaluator only understands ``path_glob:`` + ``contains_ci:`` (file-content
-    checks). It does not handle ``path:``-style JSONPath assertions on env state
-    (those are evaluated host-side by ``StateChecker.check_jsonpaths``). An
-    assertion missing ``path_glob:`` is treated as **failed** with an actionable
-    reason, so a misrouted assertion is named rather than vanishing from grading.
-
-    Each check has:
-    - path_glob: glob pattern for files (e.g., "/env/fs/agent-visible/submissions/*")
-    - contains_ci: case-insensitive substring to find in file content
-    - description: human-readable description of the check
-
-    Args:
-        checks: List of jsonpath check dicts from grading.yaml
-
-    Returns:
-        Tuple of (score, reasons_string) where score is fraction of checks passed.
-        Score is -1.0 (sentinel) when checks list is empty.
-    """
-    if not checks:
-        return -1.0, ""
-
-    passed = 0
-    total = len(checks)
-    reasons_parts: list[str] = []
-
-    for check in checks:
-        path_pattern = check.get("path_glob", "")
-        contains_ci = check.get("contains_ci", "")
-        description = check.get("description", f"Check: {contains_ci}")
-
-        if not path_pattern:
-            # Named rather than skipped: a skip scores as not-passed while reading
-            # as SKIP, which is how a misrouted assertion goes unnoticed.
-            other_path = check.get("path")
-            if other_path:
-                reasons_parts.append(
-                    f"FAIL: assertion uses 'path' (env-state JSONPath) but this "
-                    f"runner-side evaluator only supports 'path_glob' (file glob); "
-                    f"path-style assertions are graded host-side — {description}"
-                )
-            else:
-                reasons_parts.append(
-                    f"FAIL: assertion missing 'path_glob' (runner-side jsonpath "
-                    f"evaluator only supports path_glob + contains_ci) — {description}"
-                )
-            continue
-
-        # Translate logical /env/fs/agent-visible/ paths to the runner's
-        # actual /work/ tree (matching where the file tools and BashTool
-        # operate). This keeps grading/runtime consistent: the agent
-        # writes via write_file under /work/, and the grader reads from
-        # the same place.
-        resolved_pattern = path_pattern
-        if resolved_pattern.startswith("/env/fs/agent-visible/"):
-            resolved_pattern = "/work/" + resolved_pattern[len("/env/fs/agent-visible/") :]
-        elif resolved_pattern == "/env/fs/agent-visible":
-            resolved_pattern = "/work"
-
-        # Glob for matching files on the container filesystem
-        matching_files = glob.glob(resolved_pattern)
-
-        if not matching_files:
-            reasons_parts.append(f"FAIL: No files match {path_pattern} — {description}")
-            continue
-
-        # Check if any matching file contains the expected text
-        found = False
-        for file_path in matching_files:
-            try:
-                content = Path(file_path).read_text(encoding="utf-8", errors="replace")
-                if contains_ci.lower() in content.lower():
-                    found = True
-                    break
-            except Exception as exc:
-                logger.warning("Failed to read file %s: %s", file_path, exc)
-
-        if found:
-            passed += 1
-            reasons_parts.append(f"PASS: {description}")
-        else:
-            reasons_parts.append(f"FAIL: {description}")
-
-    score = passed / total if total > 0 else 0.0
-    reasons = "; ".join(reasons_parts)
-    return score, reasons
-
-
-def evaluate_jsonpath_state_checks(
-    checks: list[dict[str, Any]],
-    state: dict[str, Any],
-) -> tuple[float, str]:
-    """
-    Evaluate JSONPath assertions against the trial's structured DB state.
-
-    This is the state-oriented counterpart to evaluate_jsonpath_file_checks().
-    It supports the same JSONPath assertion shape used by the core native
-    grader: path + one of equals, equals_ci, contains, contains_ci.
-    """
-    if not checks:
-        return -1.0, ""
-
-    passed = 0
-    total = len(checks)
-    reasons_parts: list[str] = []
-
-    for check in checks:
-        path = check.get("path")
-        description = check.get("description", path or "JSONPath state check")
-        if not path:
-            reasons_parts.append(f"FAIL: Missing path — {description}")
-            continue
-
-        operators = [
-            ("equals", check.get("equals")),
-            ("equals_ci", check.get("equals_ci")),
-            ("contains", check.get("contains")),
-            ("contains_ci", check.get("contains_ci")),
-        ]
-        active = [(name, expected) for name, expected in operators if expected is not None]
-        if len(active) > 1:
-            reasons_parts.append(f"FAIL: Multiple operators at {path} — {description}")
-            continue
-
-        try:
-            matches = [match.value for match in parse(path).find(state)]
-        except Exception as exc:
-            reasons_parts.append(f"FAIL: Invalid JSONPath {path}: {exc} — {description}")
-            continue
-
-        if not matches:
-            reasons_parts.append(f"FAIL: Path not found {path} — {description}")
-            continue
-
-        if not active:
-            # Fail loud — mirrors the core native grader's contract: an
-            # assertion with no recognized operator (a typo'd or unsupported
-            # operator such as ``op:``/``expected:``, or a bare ``path:`` with
-            # no comparison) is an author error, not a passing existence check.
-            # Silently passing it would turn a strict-looking assertion into a
-            # no-op and diverge from host-side grading.
-            unknown_keys = sorted(
-                key for key in check if key not in {"path", "path_glob", "description"}
-            )
-            reasons_parts.append(
-                f"FAIL: no recognized operator at {path} (got keys: {unknown_keys}); "
-                f"supported operators are 'equals', 'equals_ci', 'contains', "
-                f"'contains_ci' — {description}"
-            )
-            continue
-
-        op_name, expected = active[0]
-        found = False
-        for value in matches:
-            if op_name == "equals" and value == expected:
-                found = True
-                break
-            if op_name == "equals_ci" and isinstance(value, str) and isinstance(expected, str):
-                if value.casefold() == expected.casefold():
-                    found = True
-                    break
-            if op_name == "contains" and contains(value, expected):
-                found = True
-                break
-            if op_name == "contains_ci" and contains(value, expected, ci=True):
-                found = True
-                break
-
-        if found:
-            passed += 1
-            reasons_parts.append(f"PASS: {description}")
-        else:
-            reasons_parts.append(
-                f"FAIL: {description} ({path}: expected {op_name} {expected!r}, got {matches[0]!r})"
-            )
-
-    score = passed / total if total > 0 else 0.0
-    reasons = "; ".join(reasons_parts)
-    return score, reasons
-
-
-async def _fetch_probe_rows(dsn: str, query: str) -> list[dict[str, Any]]:
-    """Connect to ``dsn`` via asyncpg, run ``query``, return rows as dicts.
-
-    Isolated so unit tests inject rows without a live database; the asyncpg
-    import is deferred so importing this module never requires the driver.
-    """
-    import asyncpg
-
-    conn = await asyncpg.connect(dsn)
-    try:
-        records = await conn.fetch(query)
-    finally:
-        await conn.close()
-    return [dict(record) for record in records]
-
-
-async def evaluate_db_probes(probes: list[dict[str, Any]]) -> tuple[float, str]:
-    """Evaluate substrate SQL probes against a task-declared postgres.
-
-    For each probe, connect to its ``dsn``, run its read-only ``query``, shape
-    the result into ``{"rows": [...], "row_count": N}``, and apply its ``expect``
-    JSONPath assertions via ``evaluate_jsonpath_state_checks``.
-
-    Two-level aggregation: a probe passes iff *every* ``expect`` assertion passes
-    (its JSONPath score is 1.0); the component score is the fraction of passing
-    probes. A connection/query failure is a FAILED probe with an actionable
-    reason (fail loud — never a silent pass). Empty list → -1.0 sentinel,
-    matching the file/state evaluators.
-    """
-    if not probes:
-        return -1.0, ""
-
-    passed = 0
-    total = len(probes)
-    reasons_parts: list[str] = []
-
-    for probe in probes:
-        name = probe.get("name", "")
-        description = probe.get("description") or name
-        expect = probe.get("expect", []) or []
-
-        try:
-            rows = await _fetch_probe_rows(probe.get("dsn", ""), probe.get("query", ""))
-        except Exception as exc:
-            reasons_parts.append(
-                f"FAIL: probe {name!r} could not query postgres: "
-                f"{type(exc).__name__}: {exc} — {description}"
-            )
-            continue
-
-        state = {"rows": rows, "row_count": len(rows)}
-        probe_score, probe_reasons = evaluate_jsonpath_state_checks(expect, state)
-        if probe_score == 1.0:
-            passed += 1
-            reasons_parts.append(f"PASS: probe {name!r} — {probe_reasons}")
-        else:
-            reasons_parts.append(f"FAIL: probe {name!r} — {probe_reasons}")
-
-    score = passed / total
-    return score, "; ".join(reasons_parts)
-
-
-def evaluate_jsonpath_checks(
-    checks: list[dict[str, Any]],
-    state: dict[str, Any] | None = None,
-) -> tuple[float, str]:
-    """
-    Evaluate mixed JSONPath checks.
-
-    Checks with path_glob remain file-content checks for backwards
-    compatibility. Checks with path are evaluated against structured DB state.
-    """
-    if not checks:
-        return -1.0, ""
-
-    file_checks: list[dict[str, Any]] = []
-    state_checks: list[dict[str, Any]] = []
-    invalid_checks: list[dict[str, Any]] = []
-
-    for check in checks:
-        if check.get("path_glob") is not None and check.get("path") is None:
-            file_checks.append(check)
-        elif check.get("path") is not None:
-            state_checks.append(check)
-        else:
-            invalid_checks.append(check)
-
-    passed = 0.0
-    total = len(checks)
-    reasons_parts: list[str] = []
-
-    if file_checks:
-        file_score, file_reasons = evaluate_jsonpath_file_checks(file_checks)
-        if file_score >= 0:
-            passed += file_score * len(file_checks)
-        if file_reasons:
-            reasons_parts.append(f"Files: {file_reasons}")
-
-    if state_checks:
-        if state is None:
-            reasons_parts.append("State: FAIL: DB state unavailable for JSONPath checks")
-        else:
-            state_score, state_reasons = evaluate_jsonpath_state_checks(state_checks, state)
-            if state_score >= 0:
-                passed += state_score * len(state_checks)
-            if state_reasons:
-                reasons_parts.append(f"State: {state_reasons}")
-
-    for check in invalid_checks:
-        description = check.get("description", "JSONPath check")
-        reasons_parts.append(f"FAIL: Missing path/path_glob — {description}")
-
-    score = passed / total if total > 0 else 0.0
-    reasons = "; ".join(reasons_parts)
-    return score, reasons
 
 
 def project_state_checks_to_runner_wire(slot_component: float | None) -> float:
